+++ web/modules/custom/qs_activity/src/Plugin/Block/FloatingActionsButtonsBlock.php
@@ -118,9 +118,8 @@ class FloatingActionsButtonsBlock extends BlockBase implements ContainerFactoryP
       $icon = 'communities-sm';
       $theme = 'danger';
       $url =
-        $this->urlGenerator->generateFromRoute('qs_activity.user.collection', [
+        $this->urlGenerator->generateFromRoute('qs_community.dashboard', [
           'community' => $community->id(),
-          'user' => $this->currentUser->id(),
         ]);
       $label = $this->t('qs_menu.links.account.communities');
 
@@ -258,6 +257,28 @@ class FloatingActionsButtonsBlock extends BlockBase implements ContainerFactoryP
       $label = $this->t('qs_photo.form.add.title');
     }
 
+    // Delete Photos.
+    if ($activity && in_array($route_name, [
+      'qs_photo.form.delete',
+    ])) {
+      // For everybody, show a button "My Photos".
+      $icon = 'trash';
+      $theme = 'danger';
+      $url = '#';
+      $label = $this->t('qs.photo.delete');
+    }
+
+    // Comment Photos.
+    if ($activity && in_array($route_name, [
+      'qs_photo.form.comments',
+    ])) {
+      // For everybody, show a button "My Photos".
+      $icon = 'pencil';
+      $theme = 'secondary';
+      $url = '#';
+      $label = $this->t('qs_photo.form.comment.title');
+    }
+
     // Welcome.
     if ($community && in_array($route_name, [
       'qs_community.welcome',
@@ -287,6 +308,8 @@ class FloatingActionsButtonsBlock extends BlockBase implements ContainerFactoryP
       'qs_photo.user.activities.collection',
       'qs_photo.user.form.manage',
       'qs_photo.form.add',
+      'qs_photo.form.comments',
+      'qs_photo.form.delete',
     ])) {
       $classes[] = 'active';
     }