+++ web/modules/custom/qs_photo/src/Form/UserManageForm.php
@@ -118,6 +118,12 @@ class UserManageForm extends FormBasic {
       '#options'       => $options,
     ];
 
+    if (empty($photos)) {
+      $form['empty'] = [
+        '#markup' => '<p class="text-center my-4">' . $this->t('qs_photos.photos_select.empty') . '</p>',
+      ];
+    }
+
     $form['actions'] = [
       '#type' => 'fieldset',
       '#theme_wrappers' => [
@@ -157,6 +163,9 @@ class UserManageForm extends FormBasic {
       ],
       '#value' => $this->t('qs_photos.photos_delete'),
     ];
+
+    // Remove unload script.
+    $form['#attached']['library'] = [];
 
     return $form;
   }