+++ web/themes/quartiers_solidaires/assets/components/molecules/calendar/date.js
@@ -1,0 +1,38 @@
+/**
+ * Overrides core/drupal.date library to force the use of jQuery UI DatePicker
+ */
+
+(function ($, Drupal) {
+  Drupal.behaviors.date = {
+    attach: function attach(context, settings) {
+      var $context = $(context);
+
+      $context.find('input[data-drupal-date-format]').once('datePicker').each(function () {
+        var $input = $(this);
+        var datepickerSettings = {};
+        var dateFormat = $input.data('drupalDateFormat');
+
+        // Force the type=text attribute to avoid issues in Chrome, this m***erf***er.
+        $input.attr('type', 'text');
+
+        datepickerSettings.dateFormat = dateFormat.replace('Y', 'yy').replace('m', 'mm').replace('d', 'dd');
+
+        if ($input.attr('min')) {
+          datepickerSettings.minDate = $input.attr('min');
+        }
+        if ($input.attr('max')) {
+          datepickerSettings.maxDate = $input.attr('max');
+        }
+        $input.datepicker(datepickerSettings);
+
+        // re-set the value because Chrome, this m***erf***er.
+        $input.val($input.attr('value'));
+      });
+    },
+    detach: function detach(context, settings, trigger) {
+      if (trigger === 'unload') {
+        $(context).find('input[data-drupal-date-format]').findOnce('datePicker').datepicker('destroy');
+      }
+    }
+  };
+})(jQuery, Drupal);