+++ web/themes/quartiers_solidaires/assets/components/molecules/form-control/form-comment.js
@@ -1,9 +1,8 @@
 const formComment = () => {
   (function ($) {
-    const $form = $('.js-comment-form');
+    const $textareas = $('.js-comment-form').find('textarea');
 
-    if ($form.length > 0) {
-      const $textareas = $form.find('textarea');
+    if ($textareas.length > 1) {
 
       // Copy the text to each textarea.
       const copyToOthers = text => {
@@ -13,7 +12,7 @@ const formComment = () => {
       }
 
       // Add a copy button after the first textarea
-      const btn = $(`<button class="btn btn-outline-invert mt-2">${Drupal.t('qs.comment.copy_to_other')}</button>`)
+      const btn = $(`<button class="btn btn-outline-invert shadow-to-bottom mt-2">${Drupal.t('qs.comment.copy_to_other')}</button>`)
         .on('click', function(e) {
           e.preventDefault();