+++ CHANGELOG.md
@@ -1,5 +1,51 @@
 # CHANGELOG
 
+## 1.0.0 (2018-02-26)
+ - fix #500 - Remove privilegies & subscriptions when deleting entities
+ - improve user supervisor dashboard #202
+ - fix click propagation in photoswipe gallery #497
+ - add some missing empty states in photos section #466
+ - change focus ring width to 2px instead of .2rem
+ - fix small shift when clicking on menu toggle button
+ - set event card link to weekly calendar instead of monthly #383
+ - setup favicon! #391
+ - add title to Community Apply form #297
+ - remove sentence in password recovery page #285
+ - set title in purple on Activity and Activiy Photos pages #320
+ - make invalid feedback readable on backgrounds #219
+ - ensure cleaning of redirects when entities are deleted #496 #470 #39
+ - send emails to user when communities approval is confirmed/declined #270
+ - close #411 - improve community dashboard return button
+ - add classes to toggle the flag and button visibility in event cards #234
+ - refactoring the subscription workflow to use Drupal Ajax form #234
+ - send email to user when subscription to event is confirmed #234
+ - fix broken image ratio in photoswipe gallery in view by month #444
+ - send mail to community managers when new user(s) request access to community #264
+ - send mail to community managers when new user(s) request access to community #264 - via communityApply form
+ - send mail to community managers when new user(s) request access to community #264 - via register form
+ - set the modal-open wrapper to `position: fixed` to remove scroll inertia on Safari #408
+ - send mail to subscribers of events when event is deleted #381
+ - send mail to activity organizers when events is deleted #381
+ - send mail to activity maintainers when events is deleted #381
+ - send mail to subscribers of events when event is updated #381
+ - send mail to activity organizers when events is updated #381
+ - send mail to activity maintainers when events is updated #381
+ - add the role in communities listing of supervisor dashboard #202
+ - flip the card back when using the browser back button #520
+ - remove confirm state text in ajax submit buttons #224
+ - fix translation extractor code and update translations
+ - add shadow to copy button #491
+ - fix empty state messages in photos collection by theme #466
+ - send mail to user when subscription to event is declined #377
+ - send mail to organizers & maintainers when subscription to event is confirmed #377
+ - fix Firefox glitch when flipping a card #255
+ - remove padding hacks to mitigate Firefox bugs... #255
+ - use jQuery UI Calendar by default on all browsers to avoid issues with date format on Chrome
+ - fix #507 - Google Autocomplete doesn't works when user chose a place & then change to a custom one
+ - fix #501 - Manage community Floating buttons
+ - fix IE specific cases #534
+ - fix Edge specific cases #566
+
 ## 0.1.1 (2018-02-21)
  - add Google Tag Manager